requirements.txt
open-clip-torch==2.22.0
scipy==1.11.3
scikit-learn==1.3.1
Pillow==10.0.1
Flask==2.3.3
pandas==2.2.3
numpy==1.26.0
torch==2.0.1
